+++ newrelic-alerts-configurator-dsl/src/main/kotlin/com/ocadotechnology/newrelic/alertsconfigurator/dsl/configuration/condition/ApmAppConditionDsl.kt
@@ -6,14 +6,11 @@ import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.t
 import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.violationclosetimer.ViolationCloseTimer
 import com.ocadotechnology.newrelic.alertsconfigurator.dsl.NewRelicConfigurationMarker
 import com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.terms.TermConfigurations
-import com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.terms.term as termBlock
-import com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.userDefined as userDefinedBlock
 
 @NewRelicConfigurationMarker
 class ApmAppConditionDsl {
     var conditionName: String? = null
     var enabled: Boolean = true
-    // TODO maybe string collections could be configured using String/Array<String>/varargs as well
     var applications: Collection<String> = mutableListOf()
     var metric: ApmAppCondition.Metric? = null
     var conditionScope: ApmAppCondition.ConditionScope? = null
@@ -25,7 +22,7 @@ class ApmAppConditionDsl {
     fun terms(block: TermConfigurations.() -> Unit) = terms.addAll(TermConfigurations().apply(block).terms)
 
     fun userDefined(block: UserDefinedConfigurationDsl.() -> Unit) {
-        userDefinedConfiguration = userDefinedBlock(block)
+        userDefinedConfiguration = userDefinedCondition(block)
     }
 }