+++ newrelic-alerts-configurator-dsl/src/main/kotlin/com/ocadotechnology/newrelic/alertsconfigurator/dsl/configuration/condition/terms/TermsConfigurationDsl.kt
@@ -1,10 +1,6 @@
 package com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.terms
 
-import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.DurationTerm
-import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.OperatorTerm
-import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.PriorityTerm
-import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.TermsConfiguration
-import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.TimeFunctionTerm
+import com.ocadotechnology.newrelic.alertsconfigurator.configuration.condition.terms.*
 import com.ocadotechnology.newrelic.alertsconfigurator.dsl.NewRelicConfigurationMarker
 
 @NewRelicConfigurationMarker
@@ -16,7 +12,7 @@ class TermsConfigurationDsl {
     var timeFunctionTerm: TimeFunctionTerm? = null
 }
 
-fun term(block: TermsConfigurationDsl.() -> Unit): TermsConfiguration {
+fun termConfiguration(block: TermsConfigurationDsl.() -> Unit): TermsConfiguration {
     val dsl = TermsConfigurationDsl()
     dsl.block()
 
@@ -33,6 +29,6 @@ fun term(block: TermsConfigurationDsl.() -> Unit): TermsConfiguration {
 class TermConfigurations {
     internal val terms: MutableList<TermsConfiguration> = mutableListOf()
 
-    fun term(block: TermsConfigurationDsl.() -> Unit) = terms.add(com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.terms.term(block))
+    fun term(block: TermsConfigurationDsl.() -> Unit) = terms.add(com.ocadotechnology.newrelic.alertsconfigurator.dsl.configuration.condition.terms.termConfiguration(block))
     operator fun TermsConfiguration.unaryPlus() = terms.add(this)
 }